Refactoring code and tests Improved documentation in README
