+++ .github/workflows/golangci-lint.yml
@@ -1,0 +1,22 @@
+name: golangci-lint
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+jobs:
+  golangci:
+    strategy:
+      matrix:
+        go-version: [1.13, 1.14, 1.15, 1.16]
+    name: lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-go@v2
+        with:
+          go-version: ${{ matrix.go-version }}
+      - run: go mod vendor
+      - name: golangci-lint
+        uses: golangci/golangci-lint-action@v2
+        with:
+          version: v1.40
