+++ go.mod
@@ -2,4 +2,7 @@ module github.com/liderman/text-generator
 
 go 1.15
 
-require github.com/patrickmn/go-cache v2.1.0+incompatible
+require (
+	github.com/patrickmn/go-cache v2.1.0+incompatible
+	github.com/stretchr/testify v1.7.0
+)
